# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ADAPT by LMJ API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      security: []
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      security: []
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      security: []
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /checkins:
    post:
      operationId: submitCheckin
      tags: [checkins]
      summary: Submit daily check-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinRequest"
      responses:
        "201":
          description: Check-in submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinResponse"
        "409":
          description: Already checked in today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Check-in window closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checkins/today:
    get:
      operationId: getTodayCheckin
      tags: [checkins]
      summary: Get today's check-in
      responses:
        "200":
          description: Today's check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinData"
        "404":
          description: No check-in today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required:
            - code
            - message
      required:
        - error
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [athlete, coach]
      required:
        - email
        - password
        - firstName
        - role
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - accessToken
        - refreshToken
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
        firstName:
          type: string
        lastName:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        weightKg:
          type: number
          nullable: true
        heightCm:
          type: integer
          nullable: true
        fitnessLevel:
          type: string
          nullable: true
        primaryGoal:
          type: string
          nullable: true
        cycleTracking:
          type: boolean
        coachId:
          type: string
          nullable: true
        inviteCode:
          type: string
          nullable: true
      required:
        - id
        - email
        - role
        - firstName
...
    CheckinRequest:
      type: object
      properties:
        sleep:
          type: integer
          minimum: 1
          maximum: 5
        energy:
          type: integer
          minimum: 1
          maximum: 5
        stress:
          type: integer
          minimum: 1
          maximum: 5
        soreness:
          type: integer
          minimum: 1
          maximum: 5
        motivation:
          type: integer
          minimum: 1
          maximum: 5
        hasPain:
          type: boolean
        painNotes:
          type: string
          nullable: true
        cyclePhase:
          type: string
          nullable: true
          enum: [menstrual, follicular, ovulatory, luteal]
      required:
        - sleep
        - energy
        - stress
        - soreness
        - motivation
    CheckinData:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
        sleep:
          type: integer
          nullable: true
        energy:
          type: integer
          nullable: true
        stress:
          type: integer
          nullable: true
        soreness:
          type: integer
          nullable: true
        motivation:
          type: integer
          nullable: true
        hasPain:
          type: boolean
        painNotes:
          type: string
          nullable: true
        cyclePhase:
          type: string
          nullable: true
        adaptScore:
          type: integer
        sessionMode:
          type: string
        createdAt:
          type: string
      required:
        - id
        - date
        - adaptScore
        - sessionMode
    CheckinResponse:
      type: object
      properties:
        checkin:
          $ref: "#/components/schemas/CheckinData"
        sessionPreview:
          type: object
          properties:
            name:
              type: string
            estimatedDurationMin:
              type: integer
              nullable: true
            exerciseCount:
              type: integer
          nullable: true
      required:
        - checkin
...